feat: use dead code, fix clippy warnings, clean up codebase
- enrich node API endpoints with live presence data (online, connected_at)
  from PresenceTracker instead of stale DB value
- handle MSG_TYPE_ERROR (0x03) in ts2021 noise handshake
- log VerifyRequest.source IP for DERP client audit trail
- rename test stub to create_invalid_initiation_message
- assert on dirty field in version test
- fix all clippy warnings across lib and tests
- remove unused secrecy dep from railscale-tka
- clean up garbled doc comments in preauth_key.rs and resolver.rs

+++ crates/railscale/src/lib.rs
@@ -495,7 +495,7 @@ fn build_protocol_router(state: &AppState) -> Router {
     if let Some(oidc_config) = &state.config.oidc {
         if oidc_config.rate_limit_per_minute > 0 {
             let replenish_interval_ms = 60_000 / oidc_config.rate_limit_per_minute as u64;
-            let burst_size = (oidc_config.rate_limit_per_minute / 6).clamp(3, 20) as u32;
+            let burst_size = (oidc_config.rate_limit_per_minute / 6).clamp(3, 20);
 
             let governor_conf = GovernorConfigBuilder::default()
                 .per_millisecond(replenish_interval_ms)
